Move ValueValidator class to controller package
